+++ 3/com.marginallyclever.robotoverlord/com/marginallyclever/ro3/node/nodes/class-use/Pose.html
@@ -2,7 +2,7 @@
 <html lang="en">
 <head>
 <!-- Generated by javadoc (17) on Tue Jan 09 17:14:49 UTC 2024 -->
-<title>Uses of Class com.marginallyclever.ro3.node.nodes.Pose (Robot Overlord 2.103.2 API)</title>
+<title>Uses of Class com.marginallyclever.ro3.node.nodes.pose.Pose (Robot Overlord 2.103.2 API)</title>
 <meta name="viewport" content="width=device-width, initial-scale=1">
 <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
 <meta name="dc.created" content="2024-01-09">
@@ -50,7 +50,7 @@ loadScripts(document, 'script');</script>
 <div class="flex-content">
 <main role="main">
 <div class="header">
-<h1 title="Uses of Class com.marginallyclever.ro3.node.nodes.Pose" class="title">Uses of Class<br>com.marginallyclever.ro3.node.nodes.Pose</h1>
+<h1 title="Uses of Class com.marginallyclever.ro3.node.nodes.pose.Pose" class="title">Uses of Class<br>com.marginallyclever.ro3.node.nodes.pose.Pose</h1>
 </div>
 <div class="caption"><span>Packages that use <a href="../Pose.html" title="class in com.marginallyclever.ro3.node.nodes">Pose</a></span></div>
 <div class="summary-table two-column-summary">
