+++ 3/com.marginallyclever.robotoverlord/com/marginallyclever/ro3/node/nodes/pose/MeshInstance.html
@@ -76,7 +76,7 @@ loadScripts(document, 'script');</script>
 </div>
 <div class="inheritance" title="Inheritance Tree"><a href="https://docs.oracle.com/en/java/javase/17/docs/api/java.base/java/lang/Object.html" title="class or interface in java.lang" class="external-link">java.lang.Object</a>
 <div class="inheritance"><a href="../../Node.html" title="class in com.marginallyclever.ro3.node">com.marginallyclever.ro3.node.Node</a>
-<div class="inheritance"><a href="../Pose.html" title="class in com.marginallyclever.ro3.node.nodes">com.marginallyclever.ro3.node.nodes.Pose</a>
+<div class="inheritance"><a href="../Pose.html" title="class in com.marginallyclever.ro3.node.nodes">com.marginallyclever.ro3.node.nodes.pose.Pose</a>
 <div class="inheritance">com.marginallyclever.ro3.node.nodes.pose.MeshInstance</div>
 </div>
 </div>
