+++ src/main/java/com/marginallyclever/ro3/apps/viewport/viewporttools/SelectedItems.java
@@ -1,7 +1,7 @@
 package com.marginallyclever.ro3.apps.viewport.viewporttools;
 
 import com.marginallyclever.convenience.helpers.MatrixHelper;
-import com.marginallyclever.ro3.node.nodes.Pose;
+import com.marginallyclever.ro3.node.nodes.pose.Pose;
 import com.marginallyclever.ro3.node.Node;
 
 import javax.vecmath.Matrix4d;
