+++ src/main/java/com/marginallyclever/ro3/apps/viewport/viewporttools/move/MoveUtils.java
@@ -9,7 +9,7 @@ import com.marginallyclever.ro3.apps.viewport.viewporttools.SelectedItems;
 import com.marginallyclever.ro3.apps.viewport.viewporttools.ViewportTool;
 import com.marginallyclever.ro3.node.Node;
 import com.marginallyclever.ro3.node.nodes.Camera;
-import com.marginallyclever.ro3.node.nodes.Pose;
+import com.marginallyclever.ro3.node.nodes.pose.Pose;
 import com.marginallyclever.ro3.apps.viewport.Viewport;
 
 import javax.vecmath.Matrix4d;
