+++ src/main/java/com/marginallyclever/ro3/apps/viewport/viewporttools/move/RotateToolMulti.java
@@ -10,6 +10,7 @@ import com.marginallyclever.ro3.node.Node;
 import com.marginallyclever.ro3.node.nodes.Camera;
 import com.marginallyclever.ro3.apps.viewport.ShaderProgram;
 import com.marginallyclever.ro3.apps.viewport.Viewport;
+import com.marginallyclever.ro3.node.nodes.pose.Pose;
 
 import javax.vecmath.Matrix4d;
 import javax.vecmath.Point3d;
@@ -19,7 +20,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 /**
- * <p>A tool to rotate {@link com.marginallyclever.ro3.node.nodes.Pose} noes in the {@link Viewport}.  It is a
+ * <p>A tool to rotate {@link Pose} noes in the {@link Viewport}.  It is a
  * combination of three {@link RotateToolOneAxis} viewporttools.</p> *
  */
 public class RotateToolMulti implements ViewportTool {
