+++ src/main/java/com/marginallyclever/ro3/node/nodes/limbsolver/LimbSolver.java
@@ -3,7 +3,7 @@ package com.marginallyclever.ro3.node.nodes.limbsolver;
 import com.marginallyclever.convenience.helpers.MatrixHelper;
 import com.marginallyclever.ro3.node.Node;
 import com.marginallyclever.ro3.node.NodePath;
-import com.marginallyclever.ro3.node.nodes.Pose;
+import com.marginallyclever.ro3.node.nodes.pose.Pose;
 import com.marginallyclever.ro3.node.nodes.pose.Limb;
 import org.json.JSONObject;
 import org.slf4j.Logger;
