+++ src/main/java/com/marginallyclever/ro3/node/nodes/pose/LookAt.java
@@ -1,9 +1,7 @@
 package com.marginallyclever.ro3.node.nodes.pose;
 
-import com.marginallyclever.convenience.PathCalculator;
 import com.marginallyclever.convenience.helpers.MatrixHelper;
 import com.marginallyclever.ro3.node.NodePath;
-import com.marginallyclever.ro3.node.nodes.Pose;
 import org.json.JSONObject;
 
 import javax.swing.*;
