+++ src/main/java/com/marginallyclever/ro3/node/nodes/pose/LookAtPanel.java
@@ -2,7 +2,6 @@ package com.marginallyclever.ro3.node.nodes.pose;
 
 import com.marginallyclever.ro3.apps.nodeselector.NodeSelector;
 import com.marginallyclever.ro3.PanelHelper;
-import com.marginallyclever.ro3.node.nodes.Pose;
 
 import javax.swing.*;
 import java.awt.*;
