+++ src/test/java/com/marginallyclever/ro3/node/NodeTest.java
@@ -1,7 +1,7 @@
 package com.marginallyclever.ro3.node;
 
 import com.marginallyclever.ro3.node.nodes.Camera;
-import com.marginallyclever.ro3.node.nodes.Pose;
+import com.marginallyclever.ro3.node.nodes.pose.Pose;
 import org.json.JSONObject;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
