+++ src/test/java/com/marginallyclever/ro3/node/nodes/PoseTest.java
@@ -1,6 +1,7 @@
 package com.marginallyclever.ro3.node.nodes;
 
 import com.marginallyclever.convenience.helpers.MatrixHelper;
+import com.marginallyclever.ro3.node.nodes.pose.Pose;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import javax.vecmath.Matrix4d;
